Use most conservative backend
We had difficulties leaving the default matplotlib backend, as on some
systems, including Travis' linux, this means Tk, which then needs to
be installed, and then needs to have X available.

We now attempt to use 'Agg'. Apparently everything will be easier when
matplotlib 2 comes along, and we can revert to the default backend then.

Relates to #1.