Fix: more general ovf reading
Currently, we take capitalisation of the data line in the ovf header into
account, however that does not seem to be well-defined in the standard. (We use
the same format that OOMMF writes but the documentation on the OOMMF website
uses a different standard.)

To make it more flexible, we can ignore the case.

Closes ubermag/help#266